fixed error message for post and put requests
